@@ -1,10 +1,16 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
+import { HomeComponent } from './components/home/home.component';
+import { LoginComponent } from './components/login/login.component';
 import { PlayerProfileComponent } from './components/player-profile/player-profile.component';
+import { AuthGuard } from './guards/auth.guards';
 
 const routes: Routes = [
-  { path: '', redirectTo: 'player-profile', pathMatch: 'full' },
-  { path: 'player-profile', component: PlayerProfileComponent }
+  { path: '', redirectTo: 'home', pathMatch: 'full' },
+  { path: '', component: HomeComponent },
+  { path: 'home', component: HomeComponent },
+  { path: 'login', component: LoginComponent },
+  { path: 'player-profile', component: PlayerProfileComponent, canActivate: [AuthGuard]}
 ];
 
 @NgModule({
