@@ -1,9 +1,15 @@
 import { HttpClient, HttpErrorResponse, HttpHeaders } from "@angular/common/http";
-import { OnInit } from "@angular/core";
-import { NgForm } from "@angular/forms";
+import { Component, OnInit } from "@angular/core";
 import { Router } from "@angular/router";
-import { LoginModel } from '../../../_interfaces/login.model';
+import { NgForm } from '@angular/forms';
+import { LoginModel } from '../../interface/login.model';
+import { AuthenticatedResponse } from '../../interface/authenticated-response.model';
 
+@Component({
+  selector: 'app-login',
+  templateUrl: './login.component.html',
+  styleUrls: ['./login.component.scss']
+})
 export class LoginComponent implements OnInit {
   invalidLogin: boolean | undefined;
   credentials: LoginModel = {username:'', password:''};
@@ -13,13 +19,15 @@ export class LoginComponent implements OnInit {
   }
   login = ( form: NgForm) => {
     if (form.valid) {
-      this.http.post<AuthenticatedResponse>("https://localhost:5001/api/auth/login", this.credentials, {
+      this.http.post<AuthenticatedResponse>("http://localhost:4300/api/auth/login", this.credentials, {
         headers: new HttpHeaders({ "Content-Type": "application/json"})
       })
       .subscribe({
         next: (response: AuthenticatedResponse) => {
           const token = response.token;
+          const refreshToken = response.refreshToken;
           localStorage.setItem("jwt", token); 
+          localStorage.setItem("refreshToken", refreshToken);
           this.invalidLogin = false; 
           this.router.navigate(["/"]);
         },
